requirements.txt
flask==3.0.0
python-dotenv==1.0.0
pygments==2.17.2
markdown==3.5.2 
openai